refactor: 내부 Response를 내부 클래스로 변경
